Fixed bugs to get all ProblemDAO tests to pass.

+++ TaskCompleter/src/main/java/com/artechnologiesllc/tc/persistence/dao/ProblemDAO.java
@@ -38,5 +38,5 @@ public interface ProblemDAO {
      * @return <code>Problem</code> instance from the database or
      *          <code>null</code> if id cannot be found.
      */
-    public Problem getProblem(Long problemid);
+    public Problem getProblem(int problemid);
 }
